Use normal list indexing (`List[Idx]`, `LIST.get`) to define `ListBytes:getOrDefault`
To avoid recursion and element-wise list matching when we index into a `ListBytes`,
the hooked function for normal lists should be used.
We are already doing this for the `ListBytes:primitiveLookup` function.

For the case of `ListBytes:getOrDefault`, this means to either rewrite to an
expression backed by the `LIST.get` function, or to the default value when the
index is out of range.